refactor createStore
